SOLR-16074: Fix read-only bug when deleting old index dirs
